[geo] Rename components folder to projections.
